requirements.txt
langgraph
pyyaml
opencv-python
requests
absl-py==2.3.1
attrs==25.3.0
audioread==3.0.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.2
cython_bbox==0.1.5
decorator==4.4.2
exceptiongroup==1.3.0
filelock==3.18.0
filterpy==1.4.5
fonttools==4.59.0
fsspec==2025.7.0
ftfy==6.3.1
grpcio==1.74.0
h5py==3.14.0
hf-xet==1.1.5
huggingface-hub==0.34.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
lap==0.5.12
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
loguru==0.7.3
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.1
networkx==3.4.2
ninja==1.11.1.4
numba==0.61.2
numpy==1.26.4 #cannot work in numpy >= 2.0.0
onnx==1.15.0
onnx-simplifier==0.4.10
open_clip_torch==3.0.0
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.1
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
pooch==1.8.2
proglog==0.1.12
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
pycocotools==2.0.10
pycparser==2.22
Pygments==2.19.2
pyparsing==3.2.3
pytest==7.2.0
pytest-split==0.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.1.0
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.15.3
sentencepiece==0.2.0
six==1.17.0
soundfile==0.13.1
soxr==0.5.0.post1
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.17
tokenizers==0.21.2
tomli==2.2.1
torch==2.1.0
torchvision==0.16.0
tqdm==4.67.1
transformers==4.54.0
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.170
ultralytics-thop==2.0.14
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
yolox @ git+https://github.com/Megvii-BaseDetection/YOLOX.git@6ddff4824372906469a7fae2dc3206c7aa4bbaee